Refactor grid update handling and add protocol encode/decode
Refactor grid update handling by creating a dedicated service and protocol package. Updated various components including WebSocket handling, Grid API, and tests accordingly.

+++ backend/draw/main.go
@@ -10,9 +10,9 @@ import (
 )
 
 type Req struct {
-	X     int `json:"x"`
-	Y     int `json:"y"`
-	Color int `json:"color"`
+	X     uint16 `json:"x"`
+	Y     uint16 `json:"y"`
+	Color uint8  `json:"color"`
 }
 
 func Run() {
